Refactor TodoModal into modular components and enhance type definitions
